finish unit tests
